Compress classification data modules to one file
