+++ src/netspresso_trainer/dataloaders/utils/misc.py
@@ -1,7 +1,9 @@
+import csv
 import json
 import re
 from itertools import repeat
 from pathlib import Path
+from typing import Dict, List, Optional, Tuple, Union
 
 import numpy as np
 
@@ -42,3 +44,29 @@ def get_detection_label(label_file: Path):
     label, boxes = target_array[:, 0], target_array[:, 1:]
     label = label[..., np.newaxis]
     return label, boxes
+
+
+def load_classification_class_map(labels_path: Optional[Union[str, Path]]):
+    # Assume the `map_or_filename` is path for csv label file
+    assert labels_path.exists(), f"Cannot locate specified class map file {labels_path}!"
+    class_map_ext = labels_path.suffix.lower()
+    assert class_map_ext == '.csv', f"Unsupported class map file extension ({class_map_ext})!"
+
+    with open(labels_path, newline='') as csvfile:
+        reader = csv.DictReader(csvfile)
+        file_to_idx = {row['image_id']: int(row['class']) for row in reader}
+
+    return file_to_idx
+
+
+def is_file_dict(image_dir: Union[Path, str], file_or_dir_to_idx):
+    image_dir = Path(image_dir)
+    candidate_name = list(file_or_dir_to_idx.keys())[0]
+    file_or_dir: Path = image_dir / candidate_name
+    if file_or_dir.exists():
+        return file_or_dir.is_file()
+
+    file_candidates = list(image_dir.glob(f"{candidate_name}.*"))
+    assert len(file_candidates) != 0, f"Unknown label format! Is there any something file like {file_or_dir} ?"
+
+    return True
